Add representation formats to API
