+++ lib/api/v3/projects/project_collection_representer.rb
@@ -36,14 +36,40 @@ require 'roar/json/hal'
 module API
   module V3
     module Projects
-      class ProjectCollectionRepresenter < ::API::Decorators::UnpaginatedCollection
+      class ProjectCollectionRepresenter <  ::API::Decorators::OffsetPaginatedCollection
         self.to_eager_load = ::API::V3::Projects::ProjectRepresenter.to_eager_load
         self.checked_permissions = ::API::V3::Projects::ProjectRepresenter.checked_permissions
 
-        def initialize(models, self_link:, current_user:)
+        def initialize(models, **options)
           super
 
           @represented = ::API::V3::Projects::ProjectEagerLoadingWrapper.wrap(represented)
+        end
+
+        links :representations do
+          [
+            representation_format_csv,
+            representation_format_xls
+          ]
+        end
+
+        def representation_format(format, mime_type:)
+          {
+            href: url_for(controller: :projects, action: :index, format: format, **query_params),
+            identifier: format,
+            type: mime_type,
+            title: I18n.t("export.format.#{format}")
+          }
+        end
+
+        def representation_format_xls
+          representation_format 'xls',
+                                mime_type: 'application/vnd.ms-excel'
+        end
+
+        def representation_format_csv
+          representation_format 'csv',
+                                mime_type: 'text/csv'
         end
       end
     end
