+++ lib/api/v3/work_packages/work_package_collection_representer.rb
@@ -211,6 +211,9 @@ module API
             representation_format_pdf_attachments,
             representation_format_pdf_description,
             representation_format_pdf_description_attachments,
+            representation_format_xls,
+            representation_format_xls_descriptions,
+            representation_format_xls_relations,
             representation_format_csv
           ]
 
@@ -264,6 +267,27 @@ module API
                                 i18n_key: 'pdf_with_descriptions_and_attachments',
                                 mime_type: 'application/pdf',
                                 url_query_extras: 'show_descriptions=true&show_attachments=true'
+        end
+
+        def representation_format_xls
+          representation_format 'xls',
+                                mime_type: 'application/vnd.ms-excel'
+        end
+
+        def representation_format_xls_descriptions
+          representation_format 'xls-with-descriptions',
+                                i18n_key: 'xls_with_descriptions',
+                                mime_type: 'application/vnd.ms-excel',
+                                format: 'xls',
+                                url_query_extras: 'show_descriptions=true'
+        end
+
+        def representation_format_xls_relations
+          representation_format 'xls-with-relations',
+                                i18n_key: 'xls_with_relations',
+                                mime_type: 'application/vnd.ms-excel',
+                                format: 'xls',
+                                url_query_extras: 'show_relations=true'
         end
 
         def representation_format_csv
